OAuth security e OpenAPI

+++ src/main/java/br/com/gft/desafio/api/resource/ClienteResource.java
@@ -6,6 +6,8 @@ import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -18,12 +20,15 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import br.com.gft.desafio.api.event.RecursoCriadoEvent;
 import br.com.gft.desafio.api.model.Cliente;
 import br.com.gft.desafio.api.repository.ClienteRepository;
 import br.com.gft.desafio.api.service.ClienteService;
+import io.swagger.annotations.Api;
 
 @RestController
 @RequestMapping("/api/clientes")
+@Api(tags = "Cliente")
 public class ClienteResource {
 	
 	@Autowired
@@ -31,6 +36,9 @@ public class ClienteResource {
 	
 	@Autowired
 	private ClienteService clienteService;
+	
+	@Autowired
+	private ApplicationEventPublisher publisher;
 	
 	@GetMapping
 	public List<Cliente> listarTodos(){
@@ -49,14 +57,20 @@ public class ClienteResource {
 	
 	@GetMapping("/nome/{nome}")
 	public List<Cliente> buscarPeloNome(@PathVariable String nome){
-		return clienteRepository.findByNomeContaining(nome);
+		if(!clienteRepository.findByNomeContaining(nome).isEmpty()) {
+			return clienteRepository.findByNomeContaining(nome);
+		}else {
+			throw new EmptyResultDataAccessException(1);
+		}
 	}
 	
 	@PostMapping
 	public ResponseEntity<Cliente> criar(@Valid @RequestBody Cliente cliente, HttpServletResponse response){
-		clienteRepository.save(cliente);
+		Cliente clienteSalvo = clienteRepository.save(cliente);
 		
-		return ResponseEntity.status(HttpStatus.CREATED).body(cliente);
+		publisher.publishEvent(new RecursoCriadoEvent(this, response, clienteSalvo.getId()));
+		
+		return ResponseEntity.status(HttpStatus.CREATED).body(clienteSalvo);
 	}
 	
 	@GetMapping("/{id}")
@@ -78,6 +92,10 @@ public class ClienteResource {
 	
 	@DeleteMapping("/{id}")
 	public void deletar(@PathVariable Long id) {
-		clienteRepository.deleteById(id);
+		if(clienteRepository.findById(id).isPresent()) {
+			clienteRepository.deleteById(id);
+		}else {
+			throw new EmptyResultDataAccessException(1);
+		}
 	}
 }
